// engemedical-frontend/lib/soc/interfaces/ICadastroPessoas.ts
export interface ICadastroPessoas {
  EMPRESA: string;
  CODIGO: string;
  NOME: string;
  EMAIL: string;
  TIPO: string;
  AVALIADOR: string;
  SOLICITANTE: string;
  EXAMINADOR: string;
  RESPONSAVEL: string;
  GERENTE_CONTAS: string;
  PROFISSIONAL_AGENDA: string;
  EMISSOR_ASO: string;
  TELEFONE_COM: string;
  ATIVO: string;
  ENDERECO: string;
  COMPLEMENTO: string;
  BAIRRO: string;
  CIDADE: string;
  UF: string;
  CEP: string;
  RG: string;
  CPF: string;
  NIT: string;
  CONSELHO_CLASSE: string;
  UF_CONSELHO: string;
  ESPECIALIDADE1: string;
  ESPECIALIDADE2: string;
  CD_USUARIO: string;
  USUARIO_ATIVO: string;
  TODAS_EMPRESAS: string;
  CNS: string;
  PAI: string;
  MAE: string;
  OCUPACAO: string;
  CARGO: string;
  FORMACAO_PROFISSIONAL: string;
  TIPO_VINCULO: string;
  ESCOLARIDADE: string;
  REGISTRO_FUNCIONAL: string;
  DATA_NASCIMENTO: string;
  DATA_CONTRATACAO: string;
  SEXO: string;
  HORARIO_INICIAL_TRABALHO: string;
  HORARIO_FINAL_TRABALHO: string;
  FREQUENTA_ESCOLA: string;
  BANCO: string;
  CONTA: string;
  AGENCIA: string;
  COMO_EXIBIR_COMPROMISSO_OUTROS_USUARIOS: string;
  COMO_EXIBIR_COMPROMISSO_OUTRAS_EMPRESAS: string;
  DATA_ULTIMO_ACESSO: string;
  USUARIO_EXTERNO_AGENDA: string;
  PARTICIPA_AGENDA: string;
  ACESSA_TODAS_AGENDAS: string;
  CODIGO_FUNCIONARIO_ACESSO: string;
  CODIGO_EMPRESA_ACESSO: string;
  LOGIN_USUARIO: string;
}